Move more calls related stuff to src/calls.c
... and also cleanup #includes
