+++ JuliaInterface/src/convert.c
@@ -1,5 +1,9 @@
 #include "convert.h"
 
+#include "JuliaInterface.h"
+
+#include <src/compiled.h>    // GAP headers
+
 // This function is used by LibGAP.jl
 jl_value_t * julia_gap(Obj obj)
 {
@@ -29,4 +33,3 @@ Obj gap_julia(jl_value_t * julia_obj)
     }
     return NewJuliaObj(julia_obj);
 }
-
